+++ src/BytePatch/JSON.hs
@@ -4,10 +4,9 @@
 module BytePatch.JSON where
 
 import           BytePatch.Core
-import           BytePatch.Pretty
+import qualified BytePatch.Pretty               as Pretty
 import           BytePatch.Pretty.HexByteString
 import           Data.Aeson
-import           GHC.Generics       (Generic)
 import           Text.Megaparsec
 import           Data.Void
 
@@ -24,34 +23,32 @@ jsonCfgCamelDrop x = defaultOptions
   { fieldLabelModifier = camelTo2 '_' . drop x
   , rejectUnknownFields = True }
 
-instance ToJSON   a => ToJSON   (MultiPatches a) where
-    toJSON     = genericToJSON     (jsonCfgCamelDrop 3)
-    toEncoding = genericToEncoding (jsonCfgCamelDrop 3)
-instance FromJSON a => FromJSON (MultiPatches a) where
-    parseJSON  = genericParseJSON  (jsonCfgCamelDrop 3)
+instance ToJSON   a => ToJSON   (Pretty.CommonMultiEdits a) where
+    toJSON     = genericToJSON     $ jsonCfgCamelDrop 4
+    toEncoding = genericToEncoding $ jsonCfgCamelDrop 4
+instance FromJSON a => FromJSON (Pretty.CommonMultiEdits a) where
+    parseJSON  = genericParseJSON  $ jsonCfgCamelDrop 4
 
-instance ToJSON   a => ToJSON   (MultiPatch a) where
-    toJSON     = genericToJSON     (jsonCfgCamelDrop 2)
-    toEncoding = genericToEncoding (jsonCfgCamelDrop 2)
-instance FromJSON a => FromJSON (MultiPatch a) where
-    parseJSON  = genericParseJSON  (jsonCfgCamelDrop 2)
+instance ToJSON   a => ToJSON   (Pretty.MultiEdit a) where
+    toJSON     = genericToJSON     $ jsonCfgCamelDrop 2
+    toEncoding = genericToEncoding $ jsonCfgCamelDrop 2
+instance FromJSON a => FromJSON (Pretty.MultiEdit a) where
+    parseJSON  = genericParseJSON  $ jsonCfgCamelDrop 2
 
-instance ToJSON   a => ToJSON   (Offset a) where
-    toJSON     = genericToJSON     (jsonCfgCamelDrop 1)
-    toEncoding = genericToEncoding (jsonCfgCamelDrop 1)
-instance FromJSON a => FromJSON (Offset a) where
-    parseJSON  = genericParseJSON  (jsonCfgCamelDrop 1)
+instance ToJSON   a => ToJSON   (Pretty.EditOffset a) where
+    toJSON     = genericToJSON     $ jsonCfgCamelDrop 2
+    toEncoding = genericToEncoding $ jsonCfgCamelDrop 2
+instance FromJSON a => FromJSON (Pretty.EditOffset a) where
+    parseJSON  = genericParseJSON  $ jsonCfgCamelDrop 2
 
-deriving instance Generic (OverwriteMeta a)
-instance ToJSON   a => ToJSON   (OverwriteMeta a) where
-    toJSON     = genericToJSON     (jsonCfgCamelDrop 2)
-    toEncoding = genericToEncoding (jsonCfgCamelDrop 2)
-instance FromJSON a => FromJSON (OverwriteMeta a) where
-    parseJSON  = genericParseJSON  (jsonCfgCamelDrop 2)
+instance ToJSON   a => ToJSON   (EditMeta a) where
+    toJSON     = genericToJSON     $ jsonCfgCamelDrop 2
+    toEncoding = genericToEncoding $ jsonCfgCamelDrop 2
+instance FromJSON a => FromJSON (EditMeta a) where
+    parseJSON  = genericParseJSON  $ jsonCfgCamelDrop 2
 
-deriving instance Generic (Overwrite a)
-instance ToJSON   a => ToJSON   (Overwrite a) where
+instance ToJSON   a => ToJSON   (Edit a) where
     toJSON     = genericToJSON     defaultOptions
     toEncoding = genericToEncoding defaultOptions
-instance FromJSON a => FromJSON (Overwrite a) where
+instance FromJSON a => FromJSON (Edit a) where
     parseJSON  = genericParseJSON  defaultOptions